RssFeed and RssOldItem, services and repository updated

+++ src/main/java/application/service/RssFeedService.java
@@ -12,10 +12,12 @@ import java.util.Optional;
 @Service
 public class RssFeedService {
     private RssFeedRepository rssFeedRepository;
+    private RssOldItemService rssOldItemService;
     private static Logger LOGGER = LoggerFactory.getLogger(RssFeedService.class);
 
-    public RssFeedService(RssFeedRepository rssFeedRepository) {
+    public RssFeedService(RssFeedRepository rssFeedRepository, RssOldItemService rssOldItemService) {
         this.rssFeedRepository = rssFeedRepository;
+        this.rssOldItemService = rssOldItemService;
     }
 
     public Optional<RssFeed> getFeed(Long id) {
@@ -38,7 +40,14 @@ public class RssFeedService {
     }
 
     public void removeFeed(Long id) {
-        LOGGER.info("Usuwanie rekordu o id: " + id);
-        rssFeedRepository.deleteById(id);
+        if (rssFeedRepository.findById(id).isPresent()) {
+            RssFeed rssFeed = rssFeedRepository.findById(id).get();
+            LOGGER.info("Usuwanie dzieci rekordu o id: " + id);
+            rssOldItemService.removeRecordsByRssFeed(rssFeed);
+            LOGGER.info("Usuwanie rekordu o id: " + id);
+            rssFeedRepository.deleteById(id);
+        } else {
+            LOGGER.info("Nie znaleziono rekordu do usunięcia o id: " + id);
+        }
     }
 }
